Fix login response
