+++ app/src/main/kotlin/io/github/wykopmobilny/ui/modules/links/linkdetails/LinkDetailsActivity.kt
@@ -38,32 +38,6 @@ class LinkDetailsActivity :
     InputToolbarListener,
     LinkCommentViewListener {
 
-    companion object {
-        const val EXTRA_LINK = "LINK_PARCEL"
-        const val EXTRA_LINK_ID = "EXTRA_LINKID"
-        const val EXTRA_COMMENT_ID = "EXTRA_COMMENT_ID"
-
-        fun createIntent(context: Context, link: Link) =
-            if (BuildConfig.DEBUG) {
-                LinkDetailsActivityV2.createIntent(context, link.id)
-            } else {
-                Intent(context, LinkDetailsActivity::class.java).apply {
-                    putExtra(EXTRA_LINK, link)
-                }
-            }
-
-        fun createIntent(context: Context, linkId: Long, commentId: Long? = null) =
-            if (BuildConfig.DEBUG) {
-                LinkDetailsActivityV2.createIntent(context, linkId = linkId, commentId = commentId)
-            } else {
-                Intent(context, LinkDetailsActivity::class.java).apply {
-                    putExtra(EXTRA_LINK_ID, linkId)
-                    putExtra(EXTRA_COMMENT_ID, commentId)
-                }
-            }
-
-    }
-
     @Inject
     lateinit var userManagerApi: UserManagerApi
 
@@ -354,5 +328,31 @@ class LinkDetailsActivity :
     override fun onDestroy() {
         super.onDestroy()
         presenter.unsubscribe()
+    }
+
+    companion object {
+        const val EXTRA_LINK = "LINK_PARCEL"
+        const val EXTRA_LINK_ID = "EXTRA_LINKID"
+        const val EXTRA_COMMENT_ID = "EXTRA_COMMENT_ID"
+
+        fun createIntent(context: Context, link: Link) =
+            if (BuildConfig.DEBUG) {
+                LinkDetailsActivityV2.createIntent(context, link.id)
+            } else {
+                Intent(context, LinkDetailsActivity::class.java).apply {
+                    putExtra(EXTRA_LINK, link)
+                }
+            }
+
+        fun createIntent(context: Context, linkId: Long, commentId: Long? = null) =
+            if (BuildConfig.DEBUG) {
+                LinkDetailsActivityV2.createIntent(context, linkId = linkId, commentId = commentId)
+            } else {
+                Intent(context, LinkDetailsActivity::class.java).apply {
+                    putExtra(EXTRA_LINK_ID, linkId)
+                    putExtra(EXTRA_COMMENT_ID, commentId)
+                }
+            }
+
     }
 }
